+++ src/main/java/com/iportfolio/readingnportalnotice/domain/consts/NoticeConst.java
@@ -6,7 +6,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class NoticeConst {
 
-    public static final int ACTIVATED = 1;
+    public static final short ACTIVATED = 1;
 
-    public static final int NON_ACTIVATED = 0;
+    public static final short NON_ACTIVATED = 0;
 }
